fix(funcao4): limit random vector generation to 20 and use unsigned long long to prevent factorial overflow

+++ funcao4.c
@@ -1,7 +1,7 @@
 #include "funcao4.h"
 
-int funcao4_assimetricos(int n, int *vetor) {
-    int somatorio = 0;
+unsigned long long funcao4_assimetricos(int n, int *vetor) {
+    unsigned long long somatorio = 0;
 
     for (int i = 0; i < n; i++) {
         if (vetor[i] % 2 == 0) {
@@ -9,7 +9,7 @@ int funcao4_assimetricos(int n, int *vetor) {
             somatorio += vetor[i];
         } else {
             // Caminho caro: calcula o fatorial de vetor[i]
-            int fatorial = 1;
+            unsigned long long fatorial = 1;
             for (int j = 2; j <= vetor[i]; j++) {
                 fatorial *= j;
             }
