+++ funcao4.h
@@ -1,6 +1,6 @@
 #ifndef FUNCAO4_H
 #define FUNCAO4_H
 
-int funcao4_assimetricos(int n, int *vetor);
+unsigned long long funcao4_assimetricos(int n, int *vetor);
 
 #endif
